Clean up code

+++ src/periodset.c
@@ -200,7 +200,7 @@ periodset_in(PG_FUNCTION_ARGS)
 char *
 periodset_to_string(PeriodSet *ps)
 {
-	char **strings = palloc((int) (sizeof(char *) * ps->count));
+	char **strings = palloc(sizeof(char *) * ps->count);
 	size_t outlen = 0;
 
 	for (int i = 0; i < ps->count; i++)
